Annual Report 2012: Implemented fixes suggested by various people
I think these fixes are contributions of Philip Withnal,
Ekaterina Gerasimova, David King, and Tobias Mueller.

The title needs to be capitalized and the title image looks better if
it's not a bitmap; especially the GNOME logo.

The Bug statistics page needs some love. The graphs are not representing
the values, i.e. the top lines in the four graphs are wrong (bar the
patches one).

I couldn't find the top table which should make the time go from left to right.
The asterisk on the bottom cannot be found in the table itself. So remove that.
A comma as the thousand separator will probably look better.

And there is still the donors list which, according to the title, is from 2013.
One should probably replace that with 2012 (if the data is correct!).